+++ packages/react-native-platform-override/src/OverrideFileRepositoryImpl.ts
@@ -6,11 +6,10 @@
  */
 
 import * as fs from 'fs';
+import * as globby from 'globby';
 import * as path from 'path';
 
 import {OverrideFileRepository} from './FileRepository';
-
-const DEFAULT_FILTER = /^.*\.(js|ts|jsx|tsx|cpp|h)$/;
 
 /**
  * Allows reading phsyical override files based on a passed in directory
@@ -18,50 +17,30 @@ const DEFAULT_FILTER = /^.*\.(js|ts|jsx|tsx|cpp|h)$/;
 export default class OverrideFileRepositoryImpl
   implements OverrideFileRepository {
   private baseDir: string;
-  private filter: RegExp;
 
-  constructor(baseDir: string, filter?: RegExp) {
+  constructor(baseDir: string) {
     this.baseDir = baseDir;
-    this.filter = filter || DEFAULT_FILTER;
   }
 
-  async listFiles(): Promise<Array<string>> {
-    return (await this.listFilesRec(this.baseDir))
-      .filter(file => this.filter.test(file))
-      .map(file => path.relative(this.baseDir, file));
+  async listFiles(globs: string[] = ['**']): Promise<string[]> {
+    return await globby(globs, {
+      cwd: this.baseDir,
+      dot: true,
+      ignore: ['./overrides.json'],
+    });
   }
 
-  async getFileContents(filename: string): Promise<string | null> {
+  async getFileContents(filename: string): Promise<Buffer | null> {
     const filePath = path.join(this.baseDir, filename);
     try {
-      return (await fs.promises.readFile(filePath)).toString();
+      return await fs.promises.readFile(filePath);
     } catch {
       return null;
     }
   }
 
-  async setFileContents(filename: string, content: string) {
+  async setFileContents(filename: string, content: Buffer) {
     const filePath = path.join(this.baseDir, filename);
     return fs.promises.writeFile(filePath, content);
-  }
-
-  private async listFilesRec(file: string): Promise<Array<string>> {
-    const stat = await fs.promises.stat(file);
-    if (!stat.isDirectory()) {
-      return [file];
-    }
-
-    const ret: Array<string> = [];
-    const subfiles = await fs.promises.readdir(file);
-
-    await Promise.all(
-      subfiles.map(async subfile => {
-        const fullPath = path.join(file, subfile);
-        const childFiles = await this.listFilesRec(fullPath);
-        ret.push(...childFiles);
-      }),
-    );
-
-    return ret;
   }
 }