+++ vnext/local-cli/src/generate-windows.ts
@@ -32,6 +32,8 @@ export interface GenerateOptions {
   overwrite: boolean;
   language: 'cpp' | 'cs';
   experimentalNuGetDependency: boolean;
+  nuGetTestVersion?: string;
+  nuGetTestFeed?: string;
   useWinUI3: boolean;
   verbose: boolean;
 }
@@ -44,7 +46,7 @@ export interface GenerateOptions {
  * @param  ns         namespace for the project
  * @param  options    command line options container
  */
-export function generateWindows(
+export async function generateWindows(
   projectDir: string,
   name: string,
   ns: string,
@@ -56,7 +58,7 @@ export function generateWindows(
 
   installDependencies(options);
 
-  copyProjectTemplateAndReplace(
+  await copyProjectTemplateAndReplace(
     path.join(__dirname, '..', 'templates'),
     projectDir,
     name,
@@ -71,7 +73,7 @@ const assertStableInterface: typeof generateWindows extends (
   name: string,
   ns: string,
   options: GenerateOptions,
-) => void
+) => Promise<void>
   ? true
   : never = true;
 assertStableInterface;