+++ vnext/local-cli/src/runWindows/utils/build.ts
@@ -4,28 +4,23 @@
  * @format
  */
 
-import * as fs from 'fs';
-import * as https from 'https';
-import * as os from 'os';
 import * as path from 'path';
 
-import * as MSBuildTools from './msbuildtools';
+import MSBuildTools from './msbuildtools';
 import Version from './version';
 import {commandWithProgress, newSpinner, newError} from './commandWithProgress';
-import * as util from 'util';
 import {RunWindowsOptions, BuildConfig, BuildArch} from '../runWindowsOptions';
 import {Config} from '@react-native-community/cli-types';
 
-const existsAsync = util.promisify(fs.exists);
-
 export async function buildSolution(
+  buildTools: MSBuildTools,
   slnFile: string,
   buildType: BuildConfig,
   buildArch: BuildArch,
   msBuildProps: Record<string, string>,
   verbose: boolean,
-  target: string,
-  buildLogDirectory: string,
+  target?: string,
+  buildLogDirectory?: string,
 ) {
   const minVersion = new Version(10, 0, 18362, 0);
   const allVersions = MSBuildTools.getAllAvailableUAPVersions();
@@ -35,8 +30,7 @@ export async function buildSolution(
     );
   }
 
-  const msBuildTools = MSBuildTools.findAvailableVersion(buildArch, verbose);
-  await msBuildTools.buildProject(
+  await buildTools.buildProject(
     slnFile,
     buildType,
     buildArch,
@@ -47,19 +41,17 @@ export async function buildSolution(
   );
 }
 
-async function nugetRestore(
-  nugetPath: string,
+export async function restoreNuGetPackages(
   slnFile: string,
+  buildTools: MSBuildTools,
   verbose: boolean,
-  msbuildVersion: string,
 ) {
   const text = 'Restoring NuGet packages ';
   const spinner = newSpinner(text);
-  console.log(nugetPath);
   await commandWithProgress(
     spinner,
     text,
-    nugetPath,
+    require.resolve('nuget-exe'),
     [
       'restore',
       `${slnFile}`,
@@ -67,35 +59,9 @@ async function nugetRestore(
       '-Verbosity',
       verbose ? 'normal' : 'quiet',
       '-MSBuildVersion',
-      msbuildVersion,
+      buildTools.installationVersion,
     ],
     verbose,
-  );
-}
-
-export async function restoreNuGetPackages(
-  options: RunWindowsOptions,
-  slnFile: string,
-  verbose: boolean,
-) {
-  const nugetPath = path.join(os.tmpdir(), 'nuget.4.9.2.exe');
-
-  if (!(await existsAsync(nugetPath))) {
-    const spinner = newSpinner('Downloading NuGet Binary');
-    await downloadFileWithRetry(
-      'https://dist.nuget.org/win-x86-commandline/v4.9.2/nuget.exe',
-      nugetPath,
-      1 /*retries*/,
-    );
-    spinner.succeed();
-  }
-
-  const msbuildTools = MSBuildTools.findAvailableVersion('x86', verbose);
-  await nugetRestore(
-    nugetPath,
-    slnFile,
-    verbose,
-    msbuildTools.installationVersion,
   );
 }
 
@@ -174,39 +140,4 @@ export function parseMsBuildProps(
     }
   }
   return result;
-}
-
-async function downloadFileWithRetry(
-  url: string,
-  dest: string,
-  retries: number,
-) {
-  for (let retryCount = 0; ; ++retryCount) {
-    try {
-      return await downloadFile(url, dest);
-    } catch (ex) {
-      if (retryCount === retries) {
-        throw ex;
-      }
-    }
-  }
-}
-
-function downloadFile(url: string, dest: string) {
-  const destFile = fs.createWriteStream(dest);
-
-  return new Promise((resolve, reject) => {
-    https
-      .get(url)
-      .on('response', res => res.pipe(destFile))
-      .on('finish', () => {
-        destFile.on('finish', () => {
-          destFile.close();
-          resolve();
-        });
-      })
-      .on('error', err => {
-        fs.unlink(dest, () => reject(err));
-      });
-  });
 }