+++ packages/react-native-win32/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @office-iss/react-native-win32
 
-This log was last generated on Wed, 01 Jul 2020 05:06:19 GMT and should not be manually modified.
+This log was last generated on Mon, 06 Jul 2020 23:11:46 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.0.0-canary.28
+
+Mon, 06 Jul 2020 23:11:46 GMT
+
+### Changes
+
+- Add Support for Binary Overrides to react-native-platform-override (ngerlem@microsoft.com)
 
 ## 0.0.0-canary.27