+++ packages/react-native-windows-init/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - react-native-windows-init
 
-This log was last generated on Wed, 01 Jul 2020 05:06:19 GMT and should not be manually modified.
+This log was last generated on Tue, 07 Jul 2020 05:05:23 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.1.19-0
+
+Tue, 07 Jul 2020 05:05:23 GMT
+
+### Changes
+
+- Enable experimental usage of Microsoft.ReactNative.Managed as a nuget package (dannyvv@microsoft.com)
+
+## 0.1.17
+
+Fri, 03 Jul 2020 05:05:03 GMT
+
+### Patches
+
+- Replace Custom Generator Prompt With Inquirer (ngerlem@microsoft.com)
 
 ## 0.1.16